Bibdk2021 alfa 0 order digital copy (#472)
* Generalise parsing of creators and contributors

* Role and link for creators and contributors in work

* Console.debug does not log in prod

* Update helper component

* Use ParsedCreatorsOrContributors component

* Add test for roles

* Update snaps

* Use graphQl fragment

* Format and spelling

* Order specific edition

* Filter away null pids

* Clean up props

* Generic automocker story file

* Simplify edition tests using automock utils

* Add name for query

* Allow sending manifestations to inferAccessTypes

* Use access fragment

* Move edition tests to own file
